requirements.txt
# MCP Server Requirements
mcp>=1.0.0
sqlalchemy>=2.0.0
pymysql>=1.0.0
python-dotenv>=1.0.0
asyncio
logging
# PDF processing
pymupdf>=1.23.0
# HTTP API wrapper
fastapi>=0.104.0
uvicorn>=0.24.0